Feature: GeoJSON conversion
Added GPX::Geojson module for a simple conversion of GeoJSON
features to GPX. Currently only supports LineString, Point,
MultiLineString, and MultiPoint GeoJSON features.

Note that the mapping of GeoJSON features to GPX features is not a
clear 1-to-1 conversion, e.g., a GeoJSON LineString could map to
either a GPX track or a GPX route. This implementation takes an
opinionated strategy and maps each LineString as a segment of a
GPX track. A MultiLineString maps to a GPX track, with each sub-
MultiLineString being a GPX segment. Points and MultiPoints are
converted to GPX waypoints.